requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2.4
django-braces==1.14.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-model-utils==4.1.1
django-multiselectfield==0.1.12
django-nested-admin==3.3.3
django-star-ratings==0.9.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
PyJWT==2.1.0
pyparsing==2.4.6
python-monkey-business==1.0.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.25.8
webencodings==0.5.1
